// apps/ui/src/app/client/model/meDto.ts
/**
 * LMStudio Wrapper API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MeDto { 
    username: string;
    role: string;
    subscription: string;
    isActivated: boolean;
    usedTokens: number;
    tokenCountResetDate: string;
    tokenLimit: number;
}